Splitting up the math 3d classes in their own files and trying to find a better name for the one file to include them, time to go into modules?
